Extract dashboard recommendation widget
