+++ tests/test-correctness-phase2.js
@@ -110,6 +110,7 @@ const pwaAppShellAssets = [
   '/js/dashboard-widgets.js',
   '/js/dashboard-widget-controls.js',
   '/js/dashboard-widget-renderers.js',
+  '/js/dashboard-recommendation-widget.js',
   '/js/recommendation-actions.js',
   '/js/chart-card-recs.js',
   '/js/category-glyphs.js',
